requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
neo4j==5.19.0

jinja2==3.1.4

python-multipart==0.0.9
pydantic==2.6.4
python-magic==0.4.27
sqlalchemy==2.0.31
bcrypt==4.1.3
asyncpg==0.29.0
itsdangerous==2.2.0
cryptography==42.0.8